+++ src/AtomUI.Controls.Shared/Data/ListCollectionViews/ListItemData.cs
@@ -1,15 +1,14 @@
 namespace AtomUI.Controls.Data;
 
-public record ListItemData : IListItemData
+public class ListItemData : IListItemData
 {
     public bool IsEnabled { get; set; } = true;
-    public bool IsSelected { get; set; }
     public object? Content { get; set; }
     public EntityKey? ItemKey { get; init; }
     public string? Group { get; init; }
 }
 
-public record GroupListItemData : ListItemData, IGroupListItemData
+public class GroupListItemData : ListItemData, IGroupListItemData
 {
     public bool IsGroupItem { get; set; }
 }