+++ src/AtomUI.Desktop.Controls/ListView/ListItemKeySelector.cs
@@ -1,0 +1,5 @@
+using AtomUI.Controls;
+
+namespace AtomUI.Desktop.Controls;
+
+public delegate EntityKey? ListItemKeySelector(object? item);